+++ src/components/form-concept/form-concept-pure.component.tsx
@@ -5,14 +5,14 @@ import { Field } from 'redux-form';
 import {
   withTranslations,
   Props as TranslationsProps
-} from '../../../providers/translations';
+} from '../../providers/translations';
 
-import Helptext from '../../../components/helptext/helptext.component';
+import Helptext from '../helptext/helptext.component';
 import ConceptTagsInputField from './concept-tags-input-field/concept-tags-input-field.component';
 import ConceptTagReadOnlyField from './concept-tags-readonly-field/concept-tags-readonly-field';
-import TagsInputFieldArray from '../../../components/fields/field-input-tags-objects/tags-input-field-array.component';
-import MultilingualField from '../../../components/multilingual-field/multilingual-field.component';
-import InputFieldReadonly from '../../../components/fields/field-input-readonly/field-input-readonly.component';
+import TagsInputFieldArray from '../fields/field-input-tags-objects/tags-input-field-array.component';
+import MultilingualField from '../multilingual-field/multilingual-field.component';
+import InputFieldReadonly from '../fields/field-input-readonly/field-input-readonly.component';
 
 interface ExternalProps {
   errors: any;
