+++ src/components/form-concept/form-concept-validations.jsx
@@ -1,4 +1,4 @@
-import { validateMinTwoChars } from '../../../validation/validation';
+import { validateMinTwoChars } from '../../validation/validation';
 
 const validate = ({ keyword }) => {
   let errors = {};
